+++ Common/src/main/java/fuzs/nightconfigfixes/NightConfigFixes.java
@@ -1,10 +1,10 @@
 package fuzs.nightconfigfixes;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 public class NightConfigFixes {
     public static final String MOD_ID = "nightconfigfixes";
     public static final String MOD_NAME = "Night Config Fixes";
-    public static final Logger LOGGER = LoggerFactory.getLogger(MOD_NAME);
+    public static final Logger LOGGER = LogManager.getLogger(MOD_NAME);
 }