+++ Forge/src/main/java/fuzs/nightconfigfixes/config/CheckedConfigFileTypeHandler.java
@@ -11,7 +11,8 @@ import com.electronwill.nightconfig.core.file.FileConfig;
 import com.electronwill.nightconfig.core.file.FileWatcher;
 import com.electronwill.nightconfig.core.io.ParsingException;
 import com.electronwill.nightconfig.core.io.WritingMode;
-import com.mojang.logging.LogUtils;
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Maps;
 import fuzs.nightconfigfixes.NightConfigFixes;
 import net.minecraftforge.fml.ModContainer;
 import net.minecraftforge.fml.ModList;
@@ -21,13 +22,15 @@ import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.loading.FMLConfig;
 import net.minecraftforge.fml.loading.FMLPaths;
 import net.minecraftforge.fml.util.ObfuscationReflectionHelper;
-import org.slf4j.Logger;
-import org.slf4j.Marker;
-import org.slf4j.MarkerFactory;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.Marker;
+import org.apache.logging.log4j.MarkerManager;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Map;
 import java.util.function.Function;
 
 /**
@@ -37,15 +40,15 @@ import java.util.function.Function;
  */
 public class CheckedConfigFileTypeHandler extends ConfigFileTypeHandler {
     static final ConfigFileTypeHandler TOML = new CheckedConfigFileTypeHandler();
-    static final Marker CONFIG = MarkerFactory.getMarker("CONFIG");
-    private static final Logger LOGGER = LogUtils.getLogger();
+    static final Marker CONFIG = MarkerManager.getMarker("CONFIG");
+    private static final Logger LOGGER = LogManager.getLogger();
     private static final Path DEFAULT_CONFIGS_PATH = FMLPaths.GAMEDIR.get().resolve(FMLConfig.defaultConfigPath());
+    public static final Map<String, Map<String, Object>> DEFAULT_CONFIG_VALUES = Maps.newConcurrentMap();
 
     public static void replaceDefaultConfigHandler() {
-        if (NightConfigFixesConfig.INSTANCE.getConfigParsingBehavior() != ConfigParsingBehavior.REPLACE_CONFIG_HANDLER) return;
+        if (!NightConfigFixesConfig.INSTANCE.<Boolean>getValue("recreateConfigsWhenParsingFails")) return;
         // use the IMixinConfigPlugin to switch this field as early as possible, couldn't think of something else that loads this early and is easily accessible by the mod
-        // also the TOML field not being final is very odd, let's just hope it stays like that
-        // otherwise go back to the original approach with wrapping the ModConfig instances which is currently disabled
+        // also the TOML field not being final is very odd, let's just hope it stays like that, otherwise use unsafe
         ObfuscationReflectionHelper.setPrivateValue(ConfigFileTypeHandler.class, null, CheckedConfigFileTypeHandler.TOML, "TOML");
     }
 
@@ -84,6 +87,8 @@ public class CheckedConfigFileTypeHandler extends ConfigFileTypeHandler {
             } catch (ParsingException ex) {
                 throw new ConfigLoadingException(c, ex);
             }
+            // Night Config Fixes: store values from default config, so we can retrieve them when correcting individual values
+            this.tryRegisterDefaultConfig(c);
             LOGGER.debug(CONFIG, "Loaded TOML config file {}", configPath);
             try {
                 FileWatcher.defaultInstance().addWatch(configPath, new ConfigWatcher(c, configData, Thread.currentThread().getContextClassLoader()));
@@ -108,6 +113,24 @@ public class CheckedConfigFileTypeHandler extends ConfigFileTypeHandler {
             FileWatcher.defaultInstance().removeWatch(configBasePath.resolve(config.getFileName()));
         } catch (RuntimeException e) {
             LOGGER.error("Failed to remove config {} from tracker!", configPath, e);
+        }
+    }
+
+    private void tryRegisterDefaultConfig(ModConfig modConfig) {
+        if (!NightConfigFixesConfig.INSTANCE.<Boolean>getValue("correctConfigValuesFromDefaultConfig")) return;
+        String fileName = modConfig.getFileName();
+        Path path = DEFAULT_CONFIGS_PATH.resolve(fileName);
+        if (Files.exists(path)) {
+            try (CommentedFileConfig config = CommentedFileConfig.of(path)) {
+                config.load();
+                Map<String, Object> values = config.valueMap();
+                if (values != null && !values.isEmpty()) {
+                    DEFAULT_CONFIG_VALUES.put(fileName.intern(), ImmutableMap.copyOf(values));
+                }
+                LOGGER.info(CONFIG, "Loaded default config values from file at path {}", path);
+            } catch (Exception ignored) {
+
+            }
         }
     }